# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * VERSUS TIERS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
  rememberMe?: boolean;
}

export interface ForgotPasswordInput {
  email: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  guest: 'guest',
  member: 'member',
  moderator: 'moderator',
  admin: 'admin',
  owner: 'owner',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  suspended: 'suspended',
  banned: 'banned',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  status?: UserStatus;
  createdAt: string;
  /** @nullable */
  lastLogin?: string | null;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  guest: 'guest',
  member: 'member',
  moderator: 'moderator',
  admin: 'admin',
  owner: 'owner',
} as const;

export type UserUpdateStatus = typeof UserUpdateStatus[keyof typeof UserUpdateStatus];


export const UserUpdateStatus = {
  active: 'active',
  suspended: 'suspended',
  banned: 'banned',
} as const;

export interface UserUpdate {
  username?: string;
  email?: string;
  role?: UserUpdateRole;
  status?: UserUpdateStatus;
}

export interface UserList {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface Badge {
  id: number;
  name: string;
  icon: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface Player {
  id: number;
  minecraftUsername: string;
  /** @nullable */
  uuid?: string | null;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  countryCode?: string | null;
  /** @nullable */
  skinUrl?: string | null;
  points?: number;
  /** @nullable */
  overallTier?: string | null;
  badges?: Badge[];
  createdAt: string;
  updatedAt?: string;
}

export interface Ranking {
  id: number;
  playerId: number;
  gamemodeId: number;
  /** @nullable */
  gamemodeName?: string | null;
  tier: string;
  points: number;
  winRate?: number;
  matches?: number;
  kills?: number;
  deaths?: number;
  kdr?: number;
  /** @nullable */
  rank?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export interface PlayerDetail {
  id: number;
  minecraftUsername: string;
  /** @nullable */
  uuid?: string | null;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  countryCode?: string | null;
  /** @nullable */
  skinUrl?: string | null;
  points?: number;
  /** @nullable */
  overallTier?: string | null;
  /** @nullable */
  socialLinks?: string | null;
  badges?: Badge[];
  rankings?: Ranking[];
  createdAt: string;
  updatedAt?: string;
}

export interface PlayerInput {
  minecraftUsername: string;
  uuid?: string;
  discord?: string;
  country?: string;
  countryCode?: string;
  skinUrl?: string;
  socialLinks?: string;
}

export interface PlayerUpdate {
  minecraftUsername?: string;
  uuid?: string;
  discord?: string;
  country?: string;
  countryCode?: string;
  skinUrl?: string;
  points?: number;
  overallTier?: string;
  socialLinks?: string;
}

export interface PlayerList {
  players: Player[];
  total: number;
  page: number;
  limit: number;
}

export interface Gamemode {
  id: number;
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  description?: string | null;
  enabled: boolean;
  sortOrder: number;
  /** @nullable */
  color?: string | null;
  playerCount?: number;
  createdAt: string;
}

export interface GamemodeInput {
  name: string;
  icon?: string;
  description?: string;
  enabled?: boolean;
  sortOrder?: number;
  color?: string;
}

export interface GamemodeUpdate {
  name?: string;
  icon?: string;
  description?: string;
  enabled?: boolean;
  sortOrder?: number;
  color?: string;
}

export interface RankingInput {
  playerId: number;
  gamemodeId: number;
  tier: string;
  points: number;
  winRate?: number;
  matches?: number;
  kills?: number;
  deaths?: number;
}

export interface RankingUpdate {
  tier?: string;
  points?: number;
  winRate?: number;
  matches?: number;
  kills?: number;
  deaths?: number;
}

export interface RankingWithPlayer {
  id: number;
  player: Player;
  gamemodeId: number;
  /** @nullable */
  gamemodeName?: string | null;
  tier: string;
  points: number;
  winRate?: number;
  matches?: number;
  kills?: number;
  deaths?: number;
  kdr?: number;
  rank: number;
  updatedAt?: string;
}

export interface RankingList {
  rankings: RankingWithPlayer[];
  total: number;
  page: number;
  limit: number;
}

export interface LeaderboardEntry {
  rank: number;
  player: Player;
  totalPoints: number;
  overallTier: string;
  gamemodeCount?: number;
}

export interface BadgeInput {
  name: string;
  icon: string;
  description?: string;
  color?: string;
}

export interface AssignBadgeInput {
  badgeId: number;
}

export interface SearchResults {
  players: Player[];
  total: number;
}

export interface SiteSettings {
  siteName?: string;
  siteDescription?: string;
  discordInvite?: string;
  logoUrl?: string;
  primaryColor?: string;
  backgroundType?: string;
  announcementBannerText?: string;
  announcementBannerEnabled?: boolean;
  homepageTitle?: string;
  homepageSubtitle?: string;
}

export interface SiteSettingsUpdate {
  siteName?: string;
  siteDescription?: string;
  discordInvite?: string;
  logoUrl?: string;
  primaryColor?: string;
  backgroundType?: string;
  announcementBannerText?: string;
  announcementBannerEnabled?: boolean;
  homepageTitle?: string;
  homepageSubtitle?: string;
}

export interface Announcement {
  id: number;
  text: string;
  active: boolean;
  createdAt: string;
}

export interface AnnouncementInput {
  text: string;
  active?: boolean;
}

export interface AnnouncementUpdate {
  text?: string;
  active?: boolean;
}

export interface SiteStats {
  totalUsers: number;
  totalPlayers: number;
  totalGamemodes: number;
  totalRankings: number;
  onlineUsers: number;
  recentRegistrations?: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  playerName?: string | null;
  createdAt: string;
}

export interface TierCount {
  tier: string;
  count: number;
}

export type ListUsersParams = {
page?: number;
limit?: number;
search?: string;
role?: string;
};

export type ListPlayersParams = {
page?: number;
limit?: number;
search?: string;
country?: string;
};

export type ListGamemodesParams = {
includeDisabled?: boolean;
};

export type ListRankingsParams = {
gamemodeId?: number;
page?: number;
limit?: number;
tier?: string;
sortBy?: ListRankingsSortBy;
};

export type ListRankingsSortBy = typeof ListRankingsSortBy[keyof typeof ListRankingsSortBy];


export const ListRankingsSortBy = {
  points: 'points',
  winRate: 'winRate',
  matches: 'matches',
  kills: 'kills',
  tier: 'tier',
  recent: 'recent',
} as const;

export type GetLeaderboardParams = {
limit?: number;
};

export type SearchParams = {
q?: string;
country?: string;
tier?: string;
gamemode?: string;
discord?: string;
};

export type GetRecentActivityParams = {
limit?: number;
};

export type GetTierDistributionParams = {
gamemodeId?: number;
};

export type GetTopPlayersParams = {
limit?: number;
};