# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VERSUS TIERS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Users (admin)
...
  /gamemodes:
    get:
      operationId: listGamemodes
      tags: [gamemodes]
      summary: List all gamemodes
      parameters:
        - name: includeDisabled
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Gamemodes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gamemode"
    post:
      operationId: createGamemode
      tags: [gamemodes]
      summary: Create a gamemode (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GamemodeInput"
      responses:
        "201":
          description: Created gamemode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gamemode"
  /gamemodes/{id}:
    get:
      operationId: getGamemode
      tags: [gamemodes]
      summary: Get gamemode by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Gamemode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gamemode"
    patch:
      operationId: updateGamemode
      tags: [gamemodes]
      summary: Update gamemode (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GamemodeUpdate"
      responses:
        "200":
          description: Updated gamemode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gamemode"
    delete:
      operationId: deleteGamemode
      tags: [gamemodes]
      summary: Delete gamemode (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Rankings
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 3 }
        email: { type: string }
        password: { type: string, minLength: 6 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
        rememberMe: { type: boolean }
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email: { type: string }
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token: { type: string }

    # User schemas
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        role: { type: string, enum: [guest, member, moderator, admin, owner] }
        status: { type: string, enum: [active, suspended, banned] }
        createdAt: { type: string }
        lastLogin: { type: ["string", "null"] }
...
    Gamemode:
      type: object
      required: [id, name, enabled, sortOrder, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        icon: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        enabled: { type: boolean }
        sortOrder: { type: integer }
        color: { type: ["string", "null"] }
        playerCount: { type: integer }
        createdAt: { type: string }
    GamemodeInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        icon: { type: string }
        description: { type: string }
        enabled: { type: boolean }
        sortOrder: { type: integer }
        color: { type: string }
    GamemodeUpdate:
      type: object
      properties:
        name: { type: string }
        icon: { type: string }
        description: { type: string }
        enabled: { type: boolean }
        sortOrder: { type: integer }
        color: { type: string }

    # Ranking schemas
...