# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * VERSUS TIERS API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  status?: UserStatus;
  createdAt: string;
  /** @nullable */
  lastLogin?: string | null;
}